chore(types): move types to unique dir
